requirements.txt
langchain
langchain-openai
langchain-community
langgraph
streamlit
openai
python-dotenv

chromadb
langchain-chroma

requests
beautifulsoup4
newspaper3k
ddgs
pymupdf
langchain-text-splitters
sentence-transformers

arxiv
pubmed

fpdf
python-docx

pytest
tqdm

langchain-milvus
pymilvus
